requirements.txt
fastapi[standard]
groq 
instructor 
python-dotenv
together
sqlmodel